+++ lib/controllers/application.rb
@@ -1,0 +1,8 @@
+class Chitter < Sinatra::Base
+
+  get '/' do
+    @peeps = Peep.all
+    erb :index
+  end
+
+end
