+++ lib/helpers/application.rb
@@ -1,0 +1,7 @@
+module ApplicationHelpers
+
+  def current_user
+    @current_user ||=User.get (session[:user_id]) if session[:user_id]
+  end
+
+end
